Added a single page to create a simple/variable product as well as load the required data (product data, categories, attributes). If creating, prompt user to select the type of product they are creating. If editing, checks if the product is simple or variable and loads the correct form.
